refactor(api): migrate CertificationCenterInvitation model in src
