+++ api/lib/domain/models/index.js
@@ -55,7 +55,6 @@ import { CertificationAssessment } from './CertificationAssessment.js';
 import { CertificationCandidate } from './CertificationCandidate.js';
 import { CertificationCenter } from './CertificationCenter.js';
 import { CertificationCenterForAdmin } from './CertificationCenterForAdmin.js';
-import { CertificationCenterInvitation } from './CertificationCenterInvitation.js';
 import { CertificationCenterInvitedUser } from './CertificationCenterInvitedUser.js';
 import { CertificationCenterMembership } from './CertificationCenterMembership.js';
 import { CertificationChallenge } from './CertificationChallenge.js';
@@ -182,7 +181,6 @@ export {
   CertificationCandidateForSupervising,
   CertificationCenter,
   CertificationCenterForAdmin,
-  CertificationCenterInvitation,
   CertificationCenterInvitedUser,
   CertificationCenterMembership,
   CertificationChallenge,
