+++ api/src/team/infrastructure/repositories/certification-center-invitation-repository.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../../db/knex-database-connection.js';
 import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { CertificationCenterInvitation } from '../../../../lib/domain/models/CertificationCenterInvitation.js';
+import { CertificationCenterInvitation } from '../../domain/models/CertificationCenterInvitation.js';
 
 const CERTIFICATION_CENTER_INVITATIONS = 'certification-center-invitations';
