+++ api/tests/integration/domain/usecases/get-certification-center-invitation_test.js
@@ -3,8 +3,8 @@ import {
   CancelledInvitationError,
   NotFoundError,
 } from '../../../../lib/domain/errors.js';
-import { CertificationCenterInvitation } from '../../../../lib/domain/models/CertificationCenterInvitation.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
+import { CertificationCenterInvitation } from '../../../../src/team/domain/models/CertificationCenterInvitation.js';
 import { catchErr, databaseBuilder, expect } from '../../../test-helper.js';
 
 describe('Integration | API | getCertificationCenterInvitation', function () {
