+++ api/tests/unit/infrastructure/serializers/jsonapi/certification-center-invitation-serializer_test.js
@@ -1,5 +1,5 @@
-import { CertificationCenterInvitation } from '../../../../../lib/domain/models/CertificationCenterInvitation.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/certification-center-invitation-serializer.js';
+import { CertificationCenterInvitation } from '../../../../../src/team/domain/models/CertificationCenterInvitation.js';
 import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | certification-center-invitation-serializer', function () {
